+++ app/uploaders/video_image_uploader.rb
@@ -1,0 +1,48 @@
+# encoding: utf-8
+
+class VideoImageUploader < CarrierWave::Uploader::Base
+  include CarrierWave::MiniMagick
+
+  # Choose what kind of storage to use for this uploader:
+  storage :file
+  # storage :fog
+
+  # Override the directory where uploaded files will be stored.
+  # This is a sensible default for uploaders that are meant to be mounted:
+  def store_dir
+    model_param = model.id.to_s
+    File.join("videos", model_param)
+  end
+
+  # Provide a default URL as a default if there hasn't been a file uploaded:
+  # def default_url
+  #   # For Rails 3.1+ asset pipeline compatibility:
+  #   # ActionController::Base.helpers.asset_path("fallback/" + [version_name, "default.png"].compact.join('_'))
+  #
+  #   "/images/fallback/" + [version_name, "default.png"].compact.join('_')
+  # end
+
+  # Process files as they are uploaded:
+  process resize_to_limit: [1024, 1024]
+
+  # Create different versions of your uploaded files:
+  version :preview do
+    process resize_to_fill: [360, 203]
+  end
+
+  version :thumb do
+    process resize_to_fill: [120, 67]
+  end
+
+  # Add a white list of extensions which are allowed to be uploaded.
+  def extension_white_list
+    %w(jpg jpeg png)
+  end
+
+  # Override the filename of the uploaded files:
+  # Avoid using model.id or version_name here, see uploader/store.rb for details.
+  # def filename
+  #   "video#{File.extname(original_filename)}" if original_filename
+  # end
+
+end